migrationファイル修正

+++ asset/models.py
@@ -178,33 +178,33 @@ class StockFinancialInfo(models.Model):
     date = models.DateField()
     """ stock_settlement_info """
     # 有利子負債
-    interest_bearing_debt = models.IntegerField(blank=True, null=True, verbose_name="有利子負債")
+    interest_bearing_debt = models.FloatField(blank=True, null=True, verbose_name="有利子負債")
     # ROA（総資産利益率）
     roa = models.FloatField(blank=True, null=True, verbose_name="ROA")
     # ROE
     roe = models.FloatField(blank=True, null=True, verbose_name="ROE")
     # 売上高
-    sales = models.IntegerField(blank=True, null=True, verbose_name="売上高")
+    sales = models.FloatField(blank=True, null=True, verbose_name="売上高")
     # 総資産
-    assets = models.IntegerField(blank=True, null=True, verbose_name="総資産")
+    assets = models.FloatField(blank=True, null=True, verbose_name="総資産")
     # EPS
     eps = models.FloatField(blank=True, null=True, verbose_name="EPS")
     # 当期利益
-    net_income = models.IntegerField(blank=True, null=True, verbose_name="当期利益")
+    net_income = models.FloatField(blank=True, null=True, verbose_name="当期利益")
     # BPS
     bps = models.FloatField(blank=True, null=True, verbose_name="BPS")
     # 総資産経常利益率
     roa_2 = models.FloatField(blank=True, null=True, verbose_name="総資産経常利益率")
     # 営業利益
-    operating_income = models.IntegerField(blank=True, null=True, verbose_name="営業利益")
+    operating_income = models.FloatField(blank=True, null=True, verbose_name="営業利益")
     # 自己資本比率
     equity_ratio = models.FloatField(blank=True, null=True, verbose_name="自己資本比率")
     # 資本金
-    capital = models.IntegerField(blank=True, null=True, verbose_name="資本金")
+    capital = models.FloatField(blank=True, null=True, verbose_name="資本金")
     # 経常利益
-    recurring_profit = models.IntegerField(blank=True, null=True, verbose_name="経常利益")
+    recurring_profit = models.FloatField(blank=True, null=True, verbose_name="経常利益")
     # 自己資本
-    equity = models.IntegerField(blank=True, null=True, verbose_name="自己資本")
+    equity = models.FloatField(blank=True, null=True, verbose_name="自己資本")
 
     """ stock_finance_info() """
     # PBR（実績）
@@ -212,7 +212,7 @@ class StockFinancialInfo(models.Model):
     # EPS（会社予想）
     eps_f = models.FloatField(blank=True, null=True, verbose_name="EPS（会社予想）")
     # 時価総額
-    market_value = models.IntegerField(blank=True, null=True, verbose_name="時価総額")
+    market_value = models.FloatField(blank=True, null=True, verbose_name="時価総額")
     # PER（会社予想）
     per_f = models.FloatField(blank=True, null=True, verbose_name = "PER（会社予想）")
     # 配当利回り（会社予想）
